requirements.txt
dj-database-url==0.4.2
Django==1.11.6
django-imagekit==4.0.1
django-widget-tweaks==1.4.1
Pillow==4.3.0
python-decouple==3.1